Use Redis-based counter cache for Account#statuses_count
UPDATEs on the accounts table are quite slow. Creating a status
increments the count, but the real problem is deleting statuses
in batches. UPDATEs for every deleted status bring everything
to a halt. This way, the increments/decrements happen in Redis,
and then eventually are written back into the database when the
count cache expires and is re-fetched from the database.

Cleaner re-implementation of #6281